Introduce `IdempotencyChecker` in prep for using best-effort calls (#7457)

+++ backend/canisters/notifications/impl/src/updates/c2c_sync_index.rs
@@ -6,7 +6,7 @@ use notifications_canister::c2c_sync_index::{Response::*, *};
 use notifications_index_canister::NotificationsIndexEvent;
 use stable_memory_map::StableMemoryMap;
 
-#[update(guard = "caller_is_notifications_index", msgpack = true)]
+#[update(guard = "caller_is_notifications_index", msgpack = true, fallback = true)]
 #[trace]
 fn c2c_sync_index(args: Args) -> Response {
     mutate_state(|state| c2c_sync_index_impl(args, state))
@@ -14,21 +14,27 @@ fn c2c_sync_index(args: Args) -> Response {
 
 fn c2c_sync_index_impl(args: Args, state: &mut RuntimeState) -> Response {
     for event in args.events {
-        match event {
-            NotificationsIndexEvent::SubscriptionAdded(s) => {
-                state.data.subscriptions.push(s.user_id, s.subscription);
-            }
-            NotificationsIndexEvent::SubscriptionRemoved(s) => {
-                state.data.subscriptions.remove(s.user_id, &s.p256dh_key);
-            }
-            NotificationsIndexEvent::AllSubscriptionsRemoved(u) => {
-                state.data.subscriptions.remove_all(u);
-            }
-            NotificationsIndexEvent::UserBlocked(user_id, blocked) => {
-                state.data.blocked_users.insert((blocked, user_id), ());
-            }
-            NotificationsIndexEvent::UserUnblocked(user_id, unblocked) => {
-                state.data.blocked_users.remove(&(unblocked, user_id));
+        if state.data.idempotency_checker.check(
+            state.data.notifications_index_canister_id,
+            event.created_at,
+            event.idempotency_id,
+        ) {
+            match event.value {
+                NotificationsIndexEvent::SubscriptionAdded(s) => {
+                    state.data.subscriptions.push(s.user_id, s.subscription);
+                }
+                NotificationsIndexEvent::SubscriptionRemoved(s) => {
+                    state.data.subscriptions.remove(s.user_id, &s.p256dh_key);
+                }
+                NotificationsIndexEvent::AllSubscriptionsRemoved(u) => {
+                    state.data.subscriptions.remove_all(u);
+                }
+                NotificationsIndexEvent::UserBlocked(user_id, blocked) => {
+                    state.data.blocked_users.insert((blocked, user_id), ());
+                }
+                NotificationsIndexEvent::UserUnblocked(user_id, unblocked) => {
+                    state.data.blocked_users.remove(&(unblocked, user_id));
+                }
             }
         }
     }
